refactor: move geo conditional button to packages/ui
