+++ apps/main-landing/src/app/claim/claim-page-context.tsx
@@ -28,7 +28,7 @@ const ClaimPageContext = createContext<ClaimPageContextValues | undefined>(
 
 export const ClaimPageProvider = ({ children }: Properties) => {
   const [currentContent, setCurrentContent] = useState<ClaimPageContent>(
-    CLAIM_CONTENT.LETTER,
+    CLAIM_CONTENT.CLAIM,
   );
   const [vestingPlan, setVestingPlan] = useState<VestingPlan>();
   const [eligibilityData, setEligibilityData] =
