+++ apps/main-landing/src/constants.ts
@@ -9,35 +9,3 @@ export const INTERNAL_LINK = {
 export const EXTERNAL_LINK = {
   VAULT: '/vault',
 };
-
-export const restrictedCountries = [
-  'Afghanistan',
-  'Belarus',
-  'Burundi',
-  'Central African Republic',
-  'Cuba',
-  'Crimea',
-  'Democratic Republic of the Congo',
-  'Ethiopia',
-  'Haiti',
-  'Hong Kong',
-  'Iran',
-  'Iraq',
-  'Lebanon',
-  'Libya',
-  'Mali',
-  'Myanmar',
-  'Nicaragua',
-  'North Korea',
-  'Pakistan',
-  'Russia',
-  'Somalia',
-  'South Sudan',
-  'Sudan',
-  'Syria',
-  'Ukraine',
-  'United States',
-  'Venezuela',
-  'Yemen',
-  'Zimbabwe',
-];
