+++ packages/ui/package.json
@@ -30,7 +30,8 @@
     "./radio-group": "./src/components/radio-group/index.ts",
     "./collapsible": "./src/components/collapsible/index.ts",
     "./tabs": "./src/components/tabs/index.ts",
-    "./steps": "./src/components/steps/index.ts"
+    "./steps": "./src/components/steps/index.ts",
+    "./geo-conditional-button": "./src/components/geo-conditional-button/index.ts"
   },
   "scripts": {
     "format": "prettier --check --config ../../prettier.config.mjs --ignore-path .prettierignore .",
