+++ packages/ui/src/components/geo-conditional-button/constants.ts
@@ -1,0 +1,31 @@
+export const restrictedCountries = [
+  'Afghanistan',
+  'Belarus',
+  'Burundi',
+  'Central African Republic',
+  'Cuba',
+  'Crimea',
+  'Democratic Republic of the Congo',
+  'Ethiopia',
+  'Haiti',
+  'Hong Kong',
+  'Iran',
+  'Iraq',
+  'Lebanon',
+  'Libya',
+  'Mali',
+  'Myanmar',
+  'Nicaragua',
+  'North Korea',
+  'Pakistan',
+  'Russia',
+  'Somalia',
+  'South Sudan',
+  'Sudan',
+  'Syria',
+  'Ukraine',
+  'United States',
+  'Venezuela',
+  'Yemen',
+  'Zimbabwe',
+];
