+++ packages/ui/src/components/geo-conditional-button/geo-conditional-button.tsx
@@ -2,11 +2,9 @@
 
 import { Spinner } from '@idriss-xyz/ui/spinner';
 
-import { restrictedCountries } from '@/constants';
-
-import { useGeoLocation } from '../hooks/use-geo-location';
-
+import { useGeoLocation } from './hooks/use-geo-location';
 import { BlockedButton } from './blocked-button';
+import { restrictedCountries } from './constants';
 
 type GeoConditionalButtonProperties = {
   defaultButton: React.ReactNode | React.ReactNode[]; // Support single or multiple buttons
