+++ packages/ui/src/components/geo-conditional-button/index.ts
@@ -1,0 +1,1 @@
+export { GeoConditionalButton } from './geo-conditional-button';
